Moves BgcData source and header files to /src and /include, respectively
