+++ src/disturb/WildFire.h
@@ -9,7 +9,7 @@
 #include "../data/CohortData.h"
 #include "../data/EnvData.h"
 #include "../data/FirData.h"
-#include "../data/BgcData.h"
+#include "../../include/BgcData.h"
 #include "../data/RestartData.h"
 
 #include "../inc/errorcode.h"
