+++ src/snowsoil/Soil_Bgc.h
@@ -6,7 +6,7 @@
 #include "../data/CohortData.h"
 #include "../data/EnvData.h"
 #include "../data/FirData.h"
-#include "../data/BgcData.h"
+#include "../../include/BgcData.h"
 #include "../data/RestartData.h"
 
 #include "../lookup/CohortLookup.h"
